requirements.txt
Flask==3.1.2

Flask-CORS==4.0.0
gunicorn==21.2.0

# Database - MySQL
mysql-connector-python==8.2.0
SQLAlchemy==2.0.23

# Google Cloud & Firebase
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
firebase-admin==6.3.0

# Authentication & Security
bcrypt==4.1.2
PyJWT==2.8.0

# External APIs
sendgrid==6.11.0
googlemaps==4.10.0
stripe==7.9.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0